Make more of metis-branch's metis compile.
It still doesn't compile now, with a number of entry-points in mlib that
the HOL-related code wants to use no longer present.  I guess these
will need to be copied across from the working metis code.

My feeling is that it might also be nicer to have the non-functorised
part of mlib as a "traditional" structure, perhaps with nested
structures.  E.g.,

  signature mlib = sig
    structure Portable : sig
      val ml : string
      val pointerEqual : ...
    end
    structure Useful : sig
      exception Error of string
      ...
    end
    ...
  end