updating (#3)
* Fixed misleading messages and provided a README file

* hol-mode: add hol-print-info

I recently have the problem on various machines, that have multiple
copies of HOL installed, to figure out which version is started by hol-
mode. I did this by looking at the value of HOLDIR, the current load
paths etc in ML and the used hol and holmake executables in Lisp. To
simplify this process of looking up these values, this commit adds this
functionality to hol-mode.

* Make progress removing cheats from peg_eval_print_nt

* Reduce peg_eval_print_nt cheats down to 2 (big) cases

* Add some more floating-point theorems.

Now covers square root and fused multiply add/sub.

* Add MIPS floating-point instruction encoder.

* Remove one more cheat in peg_eval_print_nt

* hol-mode: add menu entries for showing tags and axioms

* Make OpenTheory Logging more robust to uptodate

If a polymorphic constant is not uptodate because of its type, the
constant itself may still be uptodate, in which case its name will not
have "old" markers to strip off.

* Documentation and working on automating the reflectOnFailure case

* More complete theory with all definitions named.
Defined ``[<>.` to be "NOT REFINED BY" and provided an example of a proof